Add tests for comments controller, model and views

+++ app/controllers/comments_controller.rb
@@ -2,8 +2,8 @@ class CommentsController < ApplicationController
   load_and_authorize_resource :comment, except: [:create]
   before_action :find_column!
   before_action :find_comment!, only: [:edit, :update, :destroy]
-  before_action :find_card!, except: [:new]
-  before_action :flash_clear, except: [:new, :edit]
+  before_action :find_card!
+  before_action :flash_clear, except: [:edit]
   respond_to :js
 
   def index
@@ -19,25 +19,38 @@ class CommentsController < ApplicationController
       end
       flash[:success] = "Comment was successfully created." 
     else
-      flash[:error] = @comment.errors.full_messages.join("\n") 
+      flash[:error] = @comment.errors.full_messages.join("\n")
+      render status: 422 
     end
   end
 
   def edit; end
 
   def update
-    if @comment.update(comment_params)
-      flash[:success]  = "Comment was successfully updated." 
+    if current_user == @comment.user
+      if @comment.update(comment_params)
+        flash[:success]  = "Comment was successfully updated." 
+      else
+        flash[:error] = @comment.errors.full_messages.join("\n")
+        render :nothing => true, :status => :unprocessable_entity
+      end
     else
-      flash[:error] = @comment.errors.full_messages.join("\n") 
+      flash[:error] = @comment.errors.full_messages.join("\n")
+      render status: 422
     end
   end
 
   def destroy
-    if @comment.destroy
-      flash[:success]  = "Comment was successfully deleted." 
+    if current_user == @comment.user
+      if @comment.destroy
+        flash[:success]  = "Comment was successfully deleted." 
+      else
+        flash[:error] = @comment.errors.full_messages.join("\n")
+        render :nothing => true, :status => :unprocessable_entity
+      end
     else
-      flash[:error] = @comment.errors.full_messages.join("\n") 
+      flash[:error] = @comment.errors.full_messages.join("\n")
+      render status: 422
     end
   end
